requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
celery==3.1.17
contextlib2==0.5.4
Django==1.8.3
django-celery==3.1.16
django-log-request-id==1.3.1
django-simple-captcha==0.5.3
djangorestframework==3.2.2
ecdsa==0.13
et-xmlfile==1.0.1
gevent==1.2.0
greenlet==0.4.11
gunicorn==19.6.0
jdcal==1.3
jsonpatch==1.14
jsonpointer==1.10
kombu==3.0.37
#MySQL-python==1.2.5
netaddr==0.7.18
openpyxl==2.4.1
Pillow==3.4.2
pycrypto==2.6.1
python-dateutil==2.6.0
python-memcached==1.58
pytz==2016.10
raven==5.32.0
redis==2.10.5
requests==2.7.0
robotframework==3.0
robotframework-httplibrary==0.4.2
simplejson==3.8.2
six==1.10.0
sshpubkeys==2.2.0
waitress==1.0.1
WebOb==1.7.0
WebTest==2.0.24
xlwt==1.1.2
paramiko==2.1.2
impacket==0.9.15
pyOpenSSL==16.2.0
jsonfield==2.0.1
PyMySQL==0.7.10
backports.functools_lru_cache==1.0.1
